fixed height map
Changed the height map to feel more natural. Added the texture I want to use for the graves. But I realized we were supposed to use the scripts to put the locations of objects, so I'll wait to add them. Also deleted the bullet class cause we're not going to use it.